fix(plic): fix infinite plic interrupt. This confusing bug took me about one week to locate and fix.
The cause is that printing any character to the console will trigger a
UART interrupt, which will be handled by the PLIC driver. So, if we
print anything after we claim the UART interrupt from the PLIC, the UART
will generate a new interrupt (which will be pending in the PLIC).

So, when we go back to user mode, the PLIC will generate an interrupt
again, and the interrupt handler will be called again. And the process
will print something again, leading to a new interrupt. Eventually, it
makes up an infinite loop.